SCRUM-28: refactor views; keep WPF conventions
