+++ src/OmniMouse.Core/OmniMouse.Core.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk">
+<Project Sdk="Microsoft.NET.Sdk.WindowsDesktop">
   <PropertyGroup>
     <TargetFramework>net8.0</TargetFramework>
     <RootNamespace>OmniMouse.Core</RootNamespace>
